More refactor work
